refactor: update favicon path to use relative path
The favicon path in the index.html file has been updated to use a relative path instead of an absolute path. This change ensures that the favicon is correctly loaded when the project is deployed in different environments. The previous absolute path was causing the favicon to not be displayed in certain scenarios.

This commit also includes the addition of the "src/assets" directory, which contains the "apple.png" file referenced by the favicon link.

This change was made to improve the project's code organization and ensure proper display of the favicon.