feat: Add sparkle.svg asset
A new asset, sparkle.svg, has been added to the client/src/assets directory. This SVG file is a vector graphic that can be used to enhance the visual appeal of the application. It includes a sparkly design with various shapes and lines.

This commit adds the sparkle.svg asset to provide more options for visual elements in the application.